feat(ast): impl types::Match for Expr
